Reformatting, renaming to be more clear
